+++ src/main.ts
@@ -1,22 +1,27 @@
 let form: FormGenerator;
+let jsonHandler: JSONHandler;
+let svgExport: SVGExport;
 
 function OG_main(): void {
-    form = new FormGenerator(G_settings);
+    form = new FormGenerator();
+    jsonHandler = new JSONHandler();
+    svgExport = new SVGExport();
+    new DiagramGenerator();
 }
 
 function OG_update(mode: 0 | 1 = 0): void {
     form.update(mode);
-    let grapic = new DiagramGenerator(G_settings);
+    new DiagramGenerator();
 }
 
 function OG_add(type: "Row" | "Register", row: number = 0): void {
     form.add(type, row);
-    let grapic = new DiagramGenerator(G_settings);
+    new DiagramGenerator();
 }
 
 function OG_remove(type: "Row" | "Register", row: number, register: number = 0): void {
     form.remove(type, row, register);
-    let grapic = new DiagramGenerator(G_settings);
+    new DiagramGenerator();
 }
 
 function OG_showHide(type: "Row" | "Register", row: number, register: number = 0): void {
@@ -25,5 +30,14 @@ function OG_showHide(type: "Row" | "Register", row: number, register: number = 0
 
 function OG_move(type: "Row" | "Register", from: number, to: number, row: number = 0): void {
     form.move(type, from, to, row);
-    let grapic = new DiagramGenerator(G_settings);
+    new DiagramGenerator();
+}
+
+function OG_download(type: "json" | "svg"): void {
+    if(type === "json") jsonHandler.export();
+    else svgExport.download();
+}
+
+function OG_import(): void {
+    jsonHandler.import();
 }