+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2023-05-19
+
+### Added
+
+- JSON Export/Import
+- SVG Export
+
+### Changed
+
+- Use of `G_settings` in classes
+- SVG Styles from style.css to object in diagram_generator.ts
+
 ## [0.2.0] - 2023-05-17
 
 ### Changed